+++ Classes/ELCImagePicker/ELCAssetTablePicker.m
@@ -84,8 +84,8 @@
     dispatch_sync(dispatch_get_main_queue(), ^{
         [self.tableView reloadData];
         // scroll to bottom
-        int section = [self numberOfSectionsInTableView:self.tableView] - 1;
-        int row = [self tableView:self.tableView numberOfRowsInSection:section] - 1;
+        long section = [self numberOfSectionsInTableView:self.tableView] - 1;
+        long row = [self tableView:self.tableView numberOfRowsInSection:section] - 1;
         if (section >= 0 && row >= 0) {
             NSIndexPath *ip = [NSIndexPath indexPathForRow:row
                                                  inSection:section];
@@ -148,8 +148,8 @@
 
 - (NSArray *)assetsForIndexPath:(NSIndexPath *)path
 {
-    int index = path.row * self.columns;
-    int length = MIN(self.columns, [self.elcAssets count] - index);
+    long index = path.row * self.columns;
+    long length = MIN(self.columns, [self.elcAssets count] - index);
     return [self.elcAssets subarrayWithRange:NSMakeRange(index, length)];
 }
